fix: drop the sensitivity cap, state what the blend changes
The |g| <= 1 clamp fired in essentially every non-degenerate case (the
model gives 1.27 at 180 degrees of arc), so 'a short arc absorbs
one-for-one' was the clamp speaking, not the derivation -- and the
overshoot is real: one translation overshoots the middle of an arc to
reduce the error at its ends.  The clamp is removed rather than
defended.

Two docstring corrections: a full annulus does NOT make the inflation
vanish (the caller turns it into an isotropic term 130 lines later), and
the net behavioral change is now stated plainly -- for |g| <= 1 the
major eigenvalue is exactly sigma^2 whatever direction g points, so with
a max(sigma_dv, sigma_du) tier gate the derived direction sets only the
minor axis; the change against the old construction is the isotropic
floor, and the magnitude only moves the major axis above 1.

Adds direct unit tests for the derived math (the four limits, the exact
g = n identity on the rank-1 path, the overshoot, the blend's
eigenstructure and PSD-ness) rather than covering it only through
composite navigate() tests, and splits the technique module at the
1000-line cap.
