+++ tests/spindoctor/nav_technique/test_nav_technique_ring_edge.py
@@ -18,10 +18,10 @@ from spindoctor.nav_orchestrator.nav_context import NavContext
 from spindoctor.nav_technique.diagnostics import RingEdgeDiagnostics
 from spindoctor.nav_technique.dt_fitting import CoarseSearchResult
 from spindoctor.nav_technique.nav_technique_ring_edge import (
-    _RANK1_NULL_RELATIVE_THRESHOLD,
     RingEdgeNav,
     aggregate_edge_normal_angle_deg,
 )
+from spindoctor.nav_technique.ring_edge_geometry import _RANK1_NULL_RELATIVE_THRESHOLD
 
 
 def test_ring_edge_nav_recovers_planted_offset_curved(
@@ -883,7 +883,14 @@ def test_ring_edge_orbit_sigma_widens_partial_arc_along_its_radial_axis(
     result_declared = technique.navigate([declared], context)
     diff = np.asarray(result_declared.covariance_px2) - np.asarray(result_plain.covariance_px2)
     eigvals = np.linalg.eigvalsh(diff)
-    assert float(eigvals.max()) == pytest.approx(4.0, rel=1.0e-6)
+    # The arc absorbs the displacement essentially one-for-one along its own
+    # radial axis, and slightly MORE than one-for-one: a single translation
+    # overshoots the middle of an arc to reduce the error at its ends, so the
+    # derived sensitivity runs a little above 1 and the major eigenvalue a
+    # little above sigma_orbit**2 = 4.0.
+    assert float(eigvals.max()) == pytest.approx(4.22, rel=0.02)
+    # With the sensitivity at or above 1 the isotropic complement is zero, so
+    # the perpendicular axis takes nothing.
     assert float(eigvals.min()) == pytest.approx(0.0, abs=1.0e-9)
 
 
@@ -1038,7 +1045,7 @@ def test_effective_orbit_sigma_is_weight_weighted_mean(
     make_ring_feature: NavFeatureFactory,
 ) -> None:
     """Two features' sigmas combine by their share of the final LM weight."""
-    from spindoctor.nav_technique.nav_technique_ring_edge import _effective_orbit_sigma_px
+    from spindoctor.nav_technique.ring_edge_geometry import _effective_orbit_sigma_px
 
     vertices_a, normals_a = circle_polyline((50.0, 50.0), 20.0, 10)
     vertices_b, normals_b = circle_polyline((50.0, 50.0), 30.0, 10)
@@ -1067,7 +1074,7 @@ def test_effective_orbit_sigma_zero_weights_returns_max(
     make_ring_feature: NavFeatureFactory,
 ) -> None:
     """A degenerate (all-zero-weight) fit falls back to the conservative max."""
-    from spindoctor.nav_technique.nav_technique_ring_edge import _effective_orbit_sigma_px
+    from spindoctor.nav_technique.ring_edge_geometry import _effective_orbit_sigma_px
 
     vertices, normals = circle_polyline((50.0, 50.0), 20.0, 10)
     feat = make_ring_feature(
@@ -1085,7 +1092,7 @@ def test_effective_orbit_sigma_zero_when_undeclared(
     circle_polyline: CirclePolylineFactory,
     make_ring_feature: NavFeatureFactory,
 ) -> None:
-    from spindoctor.nav_technique.nav_technique_ring_edge import _effective_orbit_sigma_px
+    from spindoctor.nav_technique.ring_edge_geometry import _effective_orbit_sigma_px
 
     vertices, normals = circle_polyline((50.0, 50.0), 20.0, 10)
     feat = make_ring_feature(
